Use mod_* state files only when necessary
Generally enabling a module only requires a single command to be run on
Debian hosts, so we don't really need a separate state file to
accomplish this for each module. For those modules that need some extra
configuration to enable we use separate state files.

This means we can remove all the simple mod_* state files as they will
be handled by the general case.