Cleanup vhosts/standard.tmpl
The existing template generates a pretty messy VirtualHost file. The
updated template generates ones that looks a bit nicer and generally
cleans things up a bit.

Some more work could be done to remove options like AllowEncodedSlashes
and replace them with a generic "other options" block, but we can deal
with that at a later date if it becomes necessary.