Move creating vhostdir from config.sls to init.sls
We don't always want to include the `config.sls` state file but we do
usually want to be able to create VirtualHosts, so it seems like a good
idea to ensure that the vhostdir is created once the package is
installed. This will generally only make a difference on Redhat/CentOS
hosts that don't have a vhostdir by default.