Create vhost include config when using vhosts on RedHat/CentOS
The vhost files created by the apache salt formula are created in a
config directory that is not included by Apache by default on
RedHat/CentOS hosts. We can probably assume that the user wants to use
vhosts if they include the `apache.vhosts.standard.sls` state file.
Given that assumption, we create a config file to ensure that the vhosts
get included in the Apache config.